pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "turbodesigner2-openenv"
version = "0.1.0"
description = "OpenEnv-compatible turbomachinery design environment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.110.0",
    "uvicorn>=0.30.0",
    "flask>=2.3.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
]
baseline = [
    "matplotlib>=3.7.0",
    "torch>=2.0.0",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
include-package-data = true
py-modules = ["client", "models"]

[tool.setuptools.packages.find]
include = ["server"]
